backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
pypdf
pydantic
python-multipart